requirements.txt
Flask==3.0.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
flask-cors==4.0.0
bcrypt==4.1.2
werkzeug==3.0.1
google-generativeai==0.3.2
Pillow==10.2.0

